// src/app/interfaces/cuenta-banco.interface.ts
export interface CuentaBancoInterface {
    cuenta_Bancaria?:    any;
    descripcion?:        any;
    banco?:              any;
    id_Cuenta_Bancaria?: any;
    ban_Ini?:            any;
    ban_Ini_Mes?:        any;
    cuenta?:             any;
    num_Doc?:            any;
    saldo?:              any;
    estado?:             any;
    lugar?:              any;
    ban_Ini_Dia?:        any;
    fecha_Hora?:         any;
    userName?:           any;
    m_Fecha_Hora?:       any;
    m_UserName?:         any;
    orden?:              any;
    serie_Ini?:          any;
    serie_Fin?:          any;
    moneda?:             any;
    cuenta_M?:           any;
}